+++ packages/dash-core-components/src/components/Graph.react.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
-import {contains, intersection, filter, has, isNil, type, pluck} from 'ramda';
+import {contains, filter, has, isNil, type} from 'ramda';
 /* global Plotly:true */
 
 const filterEventData = (gd, eventData, event) => {
@@ -77,18 +77,7 @@ export default class PlotlyGraph extends Component {
         if (animate && this._hasPlotted && figure.data.length === gd.data.length) {
             return Plotly.animate(id, figure, animation_options);
         } else {
-
-            let PlotMethod;
-            if (intersection(
-                pluck('type', figure.data),
-                ['candlestick', 'ohlc']).length
-            ) {
-                PlotMethod = Plotly.newPlot;
-            } else {
-                PlotMethod = Plotly.react;
-            }
-
-            return PlotMethod(id, figure.data, figure.layout, config).then(
+            return Plotly.react(id, figure.data, figure.layout, config).then(
                 () => {
                     if (!this._hasPlotted) {
                         this.bindEvents();